Fix Docker Image issue
