+++ src/integrationTest/java/uk/gov/hmcts/reform/judicialapi/elinks/PeopleIntegrationTest.java
@@ -2,9 +2,11 @@ package uk.gov.hmcts.reform.judicialapi.elinks;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.http.HttpStatus;
+import org.springframework.test.util.ReflectionTestUtils;
 import uk.gov.hmcts.reform.judicialapi.elinks.domain.ElinkDataSchedularAudit;
 import uk.gov.hmcts.reform.judicialapi.elinks.domain.ElinksResponses;
 import uk.gov.hmcts.reform.judicialapi.elinks.util.ElinksDataLoadBaseTest;
@@ -15,6 +17,11 @@ import java.util.List;
 
 import static java.util.Comparator.comparing;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.anyList;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.willDoNothing;
+import static org.mockito.Mockito.verify;
 import static org.springframework.http.HttpStatus.FORBIDDEN;
 import static org.springframework.http.HttpStatus.OK;
 import static org.springframework.http.HttpStatus.SERVICE_UNAVAILABLE;
@@ -79,6 +86,39 @@ class PeopleIntegrationTest extends ElinksDataLoadBaseTest {
         loadPeopleData(expectedHttpStatus, RESPONSE_BODY_ERROR_MSG, testDataArguments.expectedErrorMessage());
 
         verifyPeopleDataLoadAudit(testDataArguments.expectedJobStatus());
+    }
+
+    @DisplayName("Should run scheduler job and load elinks api's data when appointment id null for authorisations")
+    @Test
+    void shouldRunSchedulerJobAndLoadAllElinksApiDataWhenAppointmentIdNull() throws IOException {
+        ReflectionTestUtils.setField(publishSidamIdService, "elinkTopicPublisher", elinkTopicPublisher);
+        given(idamTokenConfigProperties.getAuthorization()).willReturn(USER_PASSWORD);
+        willDoNothing().given(elinkTopicPublisher).sendMessage(anyList(), anyString());
+
+        stubLocationApiResponse(readJsonAsString(LOCATION_API_RESPONSE_JSON), OK);
+        stubPeopleApiResponse(readJsonAsString(PEOPLE_API_NULL_APPOINTMENT_ID_RESPONSE_JSON), OK);
+        stubLeaversApiResponse(readJsonAsString(LEAVERS_API_RESPONSE_JSON), OK);
+        stubDeletedApiResponse(readJsonAsString(DELETED_API_RESPONSE_JSON), OK);
+        stubIdamResponse(readJsonAsString(IDAM_IDS_SEARCH_RESPONSE_JSON), OK);
+        stubIdamTokenResponse(OK);
+
+        runElinksDataLoadJob();
+
+        final TestDataArguments testDataArguments = getTestDataArguments();
+
+        verifyLocationData();
+
+        verifyUserProfileData(testDataArguments);
+
+        verifyUserAppointmentsData(testDataArguments);
+
+        verifyUserAuthorisationsData(testDataArguments, true);
+
+        verifyUserJudiciaryRolesData(testDataArguments.expectedRoleSize());
+
+        verifyAudit();
+
+        verify(elinkTopicPublisher).sendMessage(anyList(), anyString());
     }
 
     private void verifyExceptions(TestDataArguments testDataArguments) {
