+++ src/integrationTest/resources/application-test.yml
@@ -137,7 +137,7 @@ testing:
     idamUrl: ${http://127.0.0.1:5000/o}
 
 
-elinksApiKey: ${judicial_api_elinks_api_key:}
+elinksApiKey: ${judicial_api_elinks_api_key_previous:}
 #URL not tested , code is for developers reference
 elinksUrl: ${ELINKS_URL:http://127.0.0.1:8000}
